+++ agent/src/index.test.ts
@@ -2,7 +2,7 @@ import assert from 'node:assert'
 import { spawnSync } from 'node:child_process'
 import path from 'node:path'
 
-import { afterAll, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest'
+import { afterAll, beforeAll, beforeEach, describe, expect, it } from 'vitest'
 
 import {
     type ContextItem,
@@ -20,7 +20,6 @@ import { TestClient, asTranscriptMessage } from './TestClient'
 import { TestWorkspace } from './TestWorkspace'
 import { decodeURIs } from './decodeURIs'
 import { explainPollyError } from './explainPollyError'
-import type { Requests } from './protocol-alias'
 import { trimEndOfLine } from './trimEndOfLine'
 
 const workspace = new TestWorkspace(path.join(__dirname, '__tests__', 'example-ts'))
@@ -944,240 +943,6 @@ describe('Agent', () => {
 
         afterAll(async () => {
             await rateLimitedClient.shutdownAndExit()
-            // Long timeout because to allow Polly.js to persist HTTP recordings
-        }, 30_000)
-    })
-
-    describe('Enterprise', () => {
-        const demoEnterpriseClient = TestClient.create({
-            workspaceRootUri: workspace.rootUri,
-            name: 'enterpriseClient',
-            credentials: TESTING_CREDENTIALS.enterprise,
-            logEventMode: 'connected-instance-only',
-        })
-        // Initialize inside beforeAll so that subsequent tests are skipped if initialization fails.
-        beforeAll(async () => {
-            const serverInfo = await demoEnterpriseClient.initialize()
-
-            expect(serverInfo.authStatus?.isLoggedIn).toBeTruthy()
-            expect(serverInfo.authStatus?.username).toStrictEqual('codytesting')
-        }, 10_000)
-
-        it('chat/submitMessage', async () => {
-            const lastMessage = await demoEnterpriseClient.sendSingleMessageToNewChat('Reply with "Yes"')
-            expect(lastMessage?.text?.trim()).toStrictEqual('Yes')
-        }, 20_000)
-
-        // Skip because it consistently fails with:
-        // Error: Test timed out in 20000ms.
-        it.skip('commands/document (enterprise client)', async () => {
-            const uri = workspace.file('src', 'example.test.ts')
-            const obtained = await demoEnterpriseClient.documentCode(uri)
-            expect(obtained).toMatchInlineSnapshot(
-                `
-              "import { expect } from 'vitest'
-              import { it } from 'vitest'
-              import { describe } from 'vitest'
-
-              describe('test block', () => {
-                  it('does 1', () => {
-                      expect(true).toBe(true)
-                  })
-
-                  it('does 2', () => {
-                      expect(true).toBe(true)
-                  })
-
-                  it('does something else', () => {
-                      // This line will error due to incorrect usage of \`performance.now\`
-                      // Record the start time of the test using the Performance API
-                      const startTime = performance.now(/* CURSOR */)
-                  })
-              })
-              "
-            `
-            )
-        }, 20_000)
-
-        it('remoteRepo/list', async () => {
-            // List a repo without a query
-            let repos: Requests['remoteRepo/list'][1]
-            do {
-                repos = await demoEnterpriseClient.request('remoteRepo/list', {
-                    query: undefined,
-                    first: 10,
-                })
-            } while (repos.state.state === 'fetching')
-            expect(repos.repos).toHaveLength(10)
-
-            // Make a paginated query.
-            const secondLastRepo = repos.repos.at(-2)
-            const moreRepos = await demoEnterpriseClient.request('remoteRepo/list', {
-                query: undefined,
-                first: 2,
-                afterId: secondLastRepo?.id,
-            })
-            expect(moreRepos.repos[0].id).toBe(repos.repos.at(-1)?.id)
-
-            // Make a query.
-            const filteredRepos = await demoEnterpriseClient.request('remoteRepo/list', {
-                query: 'sourceco',
-                first: 1000,
-            })
-            expect(
-                filteredRepos.repos.find(repo => repo.name === 'github.com/sourcegraph/cody')
-            ).toBeDefined()
-        })
-
-        it('remoteRepo/has', async () => {
-            // Query a repo that does exist.
-            const codyRepoExists = await demoEnterpriseClient.request('remoteRepo/has', {
-                repoName: 'github.com/sourcegraph/cody',
-            })
-            expect(codyRepoExists.result).toBe(true)
-
-            // Query a repo that does not exist.
-            const codyForDos = await demoEnterpriseClient.request('remoteRepo/has', {
-                repoName: 'github.com/sourcegraph/cody-edlin',
-            })
-            expect(codyForDos.result).toBe(false)
-        })
-
-        afterAll(async () => {
-            const { requests } = await demoEnterpriseClient.request('testing/networkRequests', null)
-            const nonServerInstanceRequests = requests
-                .filter(({ url }) => !url.startsWith(demoEnterpriseClient.serverEndpoint))
-                .map(({ url }) => url)
-            expect(JSON.stringify(nonServerInstanceRequests)).toStrictEqual('[]')
-            await demoEnterpriseClient.shutdownAndExit()
-            // Long timeout because to allow Polly.js to persist HTTP recordings
-        }, 30_000)
-    })
-
-    // Enterprise tests are run at demo instance, which is at a recent release version.
-    // Use this section if you need to run against S2 which is released continuously.
-    describe('Enterprise - close main branch', () => {
-        const s2EnterpriseClient = TestClient.create({
-            workspaceRootUri: workspace.rootUri,
-            name: 'enterpriseMainBranchClient',
-            credentials: TESTING_CREDENTIALS.s2,
-            logEventMode: 'connected-instance-only',
-        })
-
-        // Initialize inside beforeAll so that subsequent tests are skipped if initialization fails.
-        beforeAll(async () => {
-            const serverInfo = await s2EnterpriseClient.initialize({
-                autocompleteAdvancedProvider: 'fireworks',
-            })
-
-            expect(serverInfo.authStatus?.isLoggedIn).toBeTruthy()
-            expect(serverInfo.authStatus?.username).toStrictEqual('codytesting')
-        }, 10_000)
-
-        // Disabled because `attribution/search` GraphQL does not work on S2
-        // See https://sourcegraph.slack.com/archives/C05JDP433DL/p1714017586160079
-        it.skip('attribution/found', async () => {
-            const id = await s2EnterpriseClient.request('chat/new', null)
-            const { repoNames, error } = await s2EnterpriseClient.request('attribution/search', {
-                id,
-                snippet: 'sourcegraph.Location(new URL',
-            })
-            expect(repoNames).not.empty
-            expect(error).null
-        }, 20_000)
-
-        it('attribution/not found', async () => {
-            const id = await s2EnterpriseClient.request('chat/new', null)
-            const { repoNames, error } = await s2EnterpriseClient.request('attribution/search', {
-                id,
-                snippet: 'sourcegraph.Location(new LRU',
-            })
-            expect(repoNames).empty
-            expect(error).null
-        }, 20_000)
-
-        // Use S2 instance for Cody Context Filters enterprise tests
-        describe('Cody Context Filters for enterprise', () => {
-            it('testing/ignore/overridePolicy', async () => {
-                const onChangeCallback = vi.fn()
-
-                // `sumUri` is located inside of the github.com/sourcegraph/cody repo.
-                const ignoreTest = () =>
-                    s2EnterpriseClient.request('ignore/test', { uri: sumUri.toString() })
-                s2EnterpriseClient.registerNotification('ignore/didChange', onChangeCallback)
-
-                expect(await ignoreTest()).toStrictEqual({ policy: 'use' })
-
-                await s2EnterpriseClient.request('testing/ignore/overridePolicy', {
-                    include: [{ repoNamePattern: '' }],
-                    exclude: [{ repoNamePattern: '.*sourcegraph/cody.*' }],
-                })
-
-                expect(onChangeCallback).toBeCalledTimes(1)
-                expect(await ignoreTest()).toStrictEqual({ policy: 'ignore' })
-
-                await s2EnterpriseClient.request('testing/ignore/overridePolicy', {
-                    include: [{ repoNamePattern: '' }],
-                    exclude: [{ repoNamePattern: '.*sourcegraph/sourcegraph.*' }],
-                })
-
-                expect(onChangeCallback).toBeCalledTimes(2)
-                expect(await ignoreTest()).toStrictEqual({ policy: 'use' })
-
-                await s2EnterpriseClient.request('testing/ignore/overridePolicy', {
-                    include: [{ repoNamePattern: '' }],
-                    exclude: [{ repoNamePattern: '.*sourcegraph/sourcegraph.*' }],
-                })
-
-                // onChangeCallback is not called again because filters are the same
-                expect(onChangeCallback).toBeCalledTimes(2)
-            })
-
-            // The site config `cody.contextFilters` value on sourcegraph.sourcegraph.com instance
-            // should include `sourcegraph/cody` repo for this test to pass.
-            // Skipped because of the API error:
-            //  Request to https://sourcegraph.sourcegraph.com/.api/completions/code?client-name=vscode&client-version=v1 failed with 406 Not Acceptable: ClientCodyIgnoreCompatibilityError: Cody for vscode version "v1" doesn't match version constraint ">= 1.20.0". Please upgrade your client.
-            // https://linear.app/sourcegraph/issue/CODY-2814/fix-and-re-enable-cody-context-filters-agent-integration-test
-            it.skip('autocomplete/execute (with Cody Ignore filters)', async () => {
-                // Documents to be used as context sources.
-                await s2EnterpriseClient.openFile(animalUri)
-                await s2EnterpriseClient.openFile(squirrelUri)
-
-                // Document to generate a completion from.
-                await s2EnterpriseClient.openFile(sumUri)
-
-                const { items, completionEvent } = await s2EnterpriseClient.request(
-                    'autocomplete/execute',
-                    {
-                        uri: sumUri.toString(),
-                        position: { line: 1, character: 3 },
-                        triggerKind: 'Invoke',
-                    }
-                )
-
-                expect(items.length).toBeGreaterThan(0)
-                expect(items.map(item => item.insertText)).toMatchInlineSnapshot(
-                    `
-              [
-                "   return a + b",
-              ]
-            `
-                )
-
-                // Two documents will be checked against context filters set in site-config on S2.
-                expect(
-                    completionEvent?.params.contextSummary?.retrieverStats['jaccard-similarity']
-                        .suggestedItems
-                ).toEqual(2)
-
-                s2EnterpriseClient.notify('autocomplete/completionAccepted', {
-                    completionID: items[0].id,
-                })
-            }, 10_000)
-        })
-
-        afterAll(async () => {
-            await s2EnterpriseClient.shutdownAndExit()
             // Long timeout because to allow Polly.js to persist HTTP recordings
         }, 30_000)
     })
